subdomain/requirements.txt
requests
python-dotenv
#beautifulsoup4